review: reorganize modules and rename tests
