+++ go/cmd/multiadmin/main.go
@@ -25,8 +25,8 @@ import (
 	"os"
 	"time"
 
+	"github.com/multigres/multigres/go/admin/server"
 	"github.com/multigres/multigres/go/clustermetadata/topo"
-	"github.com/multigres/multigres/go/cmd/multiadmin/internal/server"
 	"github.com/multigres/multigres/go/servenv"
 
 	"github.com/spf13/cobra"
